+++ examples/migrations/create_users_table.js
@@ -2,7 +2,7 @@
  * Example Migration: Create Users Table
  */
 
-const Migration = require('../../lib/Migrations/Migration');
+const { Migration } = require('outlet-orm');
 
 class CreateUsersTable extends Migration {
   /**